+++ sdk/nodejs/ddos/ddosBgpInstance.ts
@@ -29,7 +29,6 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-alicloud/blob/master/website/docs/r/ddosbgp_instance.html.markdown.
  */
-/** @deprecated alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance */
 export class DdosBgpInstance extends pulumi.CustomResource {
     /**
      * Get an existing DdosBgpInstance resource's state with the given name, ID, and optional extra
@@ -40,12 +39,11 @@ export class DdosBgpInstance extends pulumi.CustomResource {
      * @param state Any extra arguments used during the lookup.
      */
     public static get(name: string, id: pulumi.Input<pulumi.ID>, state?: DdosBgpInstanceState, opts?: pulumi.CustomResourceOptions): DdosBgpInstance {
-        pulumi.log.warn("DdosBgpInstance is deprecated: alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance")
         return new DdosBgpInstance(name, <any>state, { ...opts, id: id });
     }
 
     /** @internal */
-    public static readonly __pulumiType = 'alicloud:dns/ddosBgpInstance:DdosBgpInstance';
+    public static readonly __pulumiType = 'alicloud:ddos/ddosBgpInstance:DdosBgpInstance';
 
     /**
      * Returns true if the given object is an instance of DdosBgpInstance.  This is designed to work even
@@ -94,11 +92,8 @@ export class DdosBgpInstance extends pulumi.CustomResource {
      * @param args The arguments to use to populate this resource's properties.
      * @param opts A bag of options that control this resource's behavior.
      */
-    /** @deprecated alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance */
     constructor(name: string, args: DdosBgpInstanceArgs, opts?: pulumi.CustomResourceOptions)
-    /** @deprecated alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance */
     constructor(name: string, argsOrState?: DdosBgpInstanceArgs | DdosBgpInstanceState, opts?: pulumi.CustomResourceOptions) {
-        pulumi.log.warn("DdosBgpInstance is deprecated: alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance")
         let inputs: pulumi.Inputs = {};
         if (opts && opts.id) {
             const state = argsOrState as DdosBgpInstanceState | undefined;
@@ -135,6 +130,8 @@ export class DdosBgpInstance extends pulumi.CustomResource {
         if (!opts.version) {
             opts.version = utilities.getVersion();
         }
+        const aliasOpts = { aliases: [{ type: "alicloud:dns/ddosBgpInstance:DdosBgpInstance" }] };
+        opts = opts ? pulumi.mergeOptions(opts, aliasOpts) : aliasOpts;
         super(DdosBgpInstance.__pulumiType, name, inputs, opts);
     }
 }