+++ sdk/python/pulumi_alicloud/ddos/ddos_bgp_instance.py
@@ -9,7 +9,6 @@ import pulumi.runtime
 from typing import Union
 from .. import utilities, tables
 
-warnings.warn("alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance", DeprecationWarning)
 class DdosBgpInstance(pulumi.CustomResource):
     bandwidth: pulumi.Output[float]
     """
@@ -39,7 +38,6 @@ class DdosBgpInstance(pulumi.CustomResource):
     """
     Type of the instance. Valid values: Enterprise,Professional. Default to `Enterprise`  
     """
-    warnings.warn("alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance", DeprecationWarning)
     def __init__(__self__, resource_name, opts=None, bandwidth=None, base_bandwidth=None, ip_count=None, ip_type=None, name=None, period=None, type=None, __props__=None, __name__=None, __opts__=None):
         """
         Anti-DDoS Advanced instance resource. "Ddosbgp" is the short term of this product.
@@ -49,8 +47,6 @@ class DdosBgpInstance(pulumi.CustomResource):
         > **NOTE:** Available in 1.57.0+ .
 
 
-
-        Deprecated: alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance
 
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
@@ -62,7 +58,6 @@ class DdosBgpInstance(pulumi.CustomResource):
         :param pulumi.Input[float] period: The duration that you will buy Ddosbgp instance (in month). Valid values: [1~9], 12, 24, 36. Default to 12. At present, the provider does not support modify "period".
         :param pulumi.Input[str] type: Type of the instance. Valid values: Enterprise,Professional. Default to `Enterprise`  
         """
-        pulumi.log.warn("DdosBgpInstance is deprecated: alicloud.dns.DdosBgpInstance has been deprecated in favour of alicloud.ddos.DdosBgpInstance")
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
             resource_name = __name__
@@ -93,8 +88,10 @@ class DdosBgpInstance(pulumi.CustomResource):
             __props__['name'] = name
             __props__['period'] = period
             __props__['type'] = type
+        alias_opts = pulumi.ResourceOptions(aliases=[pulumi.Alias(type_="alicloud:dns/ddosBgpInstance:DdosBgpInstance")])
+        opts = pulumi.ResourceOptions.merge(opts, alias_opts)
         super(DdosBgpInstance, __self__).__init__(
-            'alicloud:dns/ddosBgpInstance:DdosBgpInstance',
+            'alicloud:ddos/ddosBgpInstance:DdosBgpInstance',
             resource_name,
             __props__,
             opts)