+++ sdk/python/pulumi_alicloud/ddos/ddos_coo_instance.py
@@ -9,7 +9,6 @@ import pulumi.runtime
 from typing import Union
 from .. import utilities, tables
 
-warnings.warn("alicloud.dns.DdosCooInstance has been deprecated in favour of alicloud.ddos.DdosCooInstance", DeprecationWarning)
 class DdosCooInstance(pulumi.CustomResource):
     bandwidth: pulumi.Output[str]
     """
@@ -39,7 +38,6 @@ class DdosCooInstance(pulumi.CustomResource):
     """
     Business bandwidth of the instance. At leaset 100. Increased 100 per step, such as 100, 200, 300. The unit is Mbps. Only support upgrade.
     """
-    warnings.warn("alicloud.dns.DdosCooInstance has been deprecated in favour of alicloud.ddos.DdosCooInstance", DeprecationWarning)
     def __init__(__self__, resource_name, opts=None, bandwidth=None, base_bandwidth=None, domain_count=None, name=None, period=None, port_count=None, service_bandwidth=None, __props__=None, __name__=None, __opts__=None):
         """
         BGP-Line Anti-DDoS instance resource. "Ddoscoo" is the short term of this product. See [What is Anti-DDoS Pro](https://www.alibabacloud.com/help/doc-detail/69319.htm).
@@ -52,8 +50,6 @@ class DdosCooInstance(pulumi.CustomResource):
 
 
 
-        Deprecated: alicloud.dns.DdosCooInstance has been deprecated in favour of alicloud.ddos.DdosCooInstance
-
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] bandwidth: Elastic defend bandwidth of the instance. This value must be larger than the base defend bandwidth. Valid values: 30, 60, 100, 300, 400, 500, 600. The unit is Gbps. Only support upgrade.
@@ -64,7 +60,6 @@ class DdosCooInstance(pulumi.CustomResource):
         :param pulumi.Input[str] port_count: Port retransmission rule count of the instance. At least 50. Increase 5 per step, such as 55, 60, 65. Only support upgrade.
         :param pulumi.Input[str] service_bandwidth: Business bandwidth of the instance. At leaset 100. Increased 100 per step, such as 100, 200, 300. The unit is Mbps. Only support upgrade.
         """
-        pulumi.log.warn("DdosCooInstance is deprecated: alicloud.dns.DdosCooInstance has been deprecated in favour of alicloud.ddos.DdosCooInstance")
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
             resource_name = __name__
@@ -99,8 +94,10 @@ class DdosCooInstance(pulumi.CustomResource):
             if service_bandwidth is None:
                 raise TypeError("Missing required property 'service_bandwidth'")
             __props__['service_bandwidth'] = service_bandwidth
+        alias_opts = pulumi.ResourceOptions(aliases=[pulumi.Alias(type_="alicloud:dns/ddosCooInstance:DdosCooInstance")])
+        opts = pulumi.ResourceOptions.merge(opts, alias_opts)
         super(DdosCooInstance, __self__).__init__(
-            'alicloud:dns/ddosCooInstance:DdosCooInstance',
+            'alicloud:ddos/ddosCooInstance:DdosCooInstance',
             resource_name,
             __props__,
             opts)